refactor(Experience): switch to POJOs
